v2.1 aiogramm

+++ BD.py
@@ -1,5 +1,8 @@
+import os
+import os.path
 import sqlite3
 from datetime import datetime
+from openpyxl import Workbook
 
 async def telegramm_base(state):
     async with state.proxy() as data:
@@ -48,25 +51,25 @@ async def del_user(a):
     conn.close()
     return b
 
-async def act_user(b):
-    tims = datetime.now()
-    sur = '---'
-    conn = sqlite3.connect('table_mint_2.db')
-    cursor = conn.cursor()
-    cursor.execute("""CREATE TABLE IF NOT EXISTS list_2
-                      (number INTEGER PRIMARY KEY, name TEXT (30), username TEXT (30), user_id INTEGER (20), 
-                      sex TEXT (10), time DATETIME, sur TEXT (3))""")
-    cursor.execute('INSERT INTO list_2 (name, username, user_id, sex, time, sur) VALUES (?,?,?,?,?,?)',
-                   (b[1], b[3], b[4], b[5], tims, sur))
-    conn.commit()
-    b = f'Пользователь {b[1]} активирован'
-    conn.close()
-    return b
+# async def act_user(b):
+#     tims = datetime.now()
+#     sur, sur_select = '---', '0'
+#     conn = sqlite3.connect('table_mint_2.db')
+#     cursor = conn.cursor()
+#     cursor.execute("""CREATE TABLE IF NOT EXISTS list_2
+#                       (number INTEGER PRIMARY KEY, name TEXT (30), username TEXT (30), user_id INTEGER (20),
+#                       sex TEXT (10), time DATETIME, sur TEXT (3), w_c INTEGER (2), w_s INTEGER (2), vin INTEGER (2))""")
+#     cursor.execute('INSERT INTO list_2 (name, username, user_id, sex, time, sur, w_c, w_s, vin) VALUES (?,?,?,?,?,?,?,?,?)',
+#                    (b[1], b[3], b[4], b[5], tims, sur, sur_select, sur_select, sur_select))
+#     conn.commit()
+#     b = f'Пользователь {b[1]} активирован'
+#     conn.close()
+#     return b
 
 async def get_info_act():
     conn = sqlite3.connect('table_mint_2.db')
     cursor = conn.cursor()
-    cursor.execute("SELECT number, name, username, user_id, sex, time, sur FROM list_2")
+    cursor.execute("SELECT number, name, username, user_id, sex, time, sur, w_c, w_s, vin FROM list_2")
     res = cursor.fetchall()
     list = []
     for i in res:
@@ -80,16 +83,36 @@ async def interval():
     conn = sqlite3.connect('table_mint_2.db')
     cursor = conn.cursor()
     t_n = datetime.now()
-    current_time = datetime.now().replace(hour=22, minute=10, second=0)
-    if t_n > current_time:
-        cursor.execute("DELETE FROM list_2")
-        conn.commit()
-        conn.close()
+    t_n1 = datetime.now().strftime("%Y-%m-%d")
+    current_time = datetime.now().replace(hour=14, minute=00, second=0)
+    current_time1 = datetime.now().replace(hour=14, minute=20, second=0)
+    if t_n > current_time and t_n < current_time1:
+        folder_name = "reports"
+        filename = f'Отчет_{t_n1}.xlsx'
+        folder_path = os.path.join(os.getcwd(), folder_name)
+        file_path = os.path.join(folder_path, filename)
+        if not os.path.isfile(filename):
+            cursor.execute('SELECT * FROM list_2')
+            data = cursor.fetchall()
+            workbook = Workbook()
+            sheet = workbook.active
+            for row_index, row_data in enumerate(data, start=1):
+                for col_index, col_data in enumerate(row_data, start=1):
+                    sheet.cell(row=row_index, column=col_index, value=col_data)
+            if not os.path.exists(folder_path):
+                os.makedirs(folder_path)
+            workbook.save(file_path)
+
+            cursor.execute("DELETE FROM list_2")
+            conn.commit()
+            conn.close()
     else:
-        cursor.execute("SELECT time, number, user_id FROM list_2 WHERE sur = ?", ("---",))
+        cursor.execute("SELECT time, number, user_id, w_c, w_s, vin FROM list_2 WHERE sur = ?", ("---",))
         res = cursor.fetchall()
         if res:
             for i in res:
+                if int(i[3]) + int(i[4]) + int(i[5]) >= 3:
+                    continue
                 time_1 = datetime.strptime(str(i[0]), "%Y-%m-%d %H:%M:%S.%f")
                 diff_minutes = t_n - time_1
                 minutes = diff_minutes.total_seconds() // 60
